refactor: split index.ts and app.ts
